// src/token-manager/HTTPClient/vehicle/protobuf/outputs/signatures.ts
export enum Tag {
  TAG_SIGNATURE_TYPE = 0,
  TAG_DOMAIN = 1,
  TAG_PERSONALIZATION = 2,
  TAG_EPOCH = 3,
  TAG_EXPIRES_AT = 4,
  TAG_COUNTER = 5,
  TAG_CHALLENGE = 6,
  TAG_FLAGS = 7,
  TAG_REQUEST_HASH = 8,
  TAG_FAULT = 9,
  TAG_END = 255,
  UNRECOGNIZED = -1,
}
// ...
export enum SignatureType {
  SIGNATURE_TYPE_AES_GCM = 0,
  SIGNATURE_TYPE_AES_GCM_PERSONALIZED = 5,
  SIGNATURE_TYPE_HMAC = 6,
  SIGNATURE_TYPE_HMAC_PERSONALIZED = 8,
  SIGNATURE_TYPE_AES_GCM_RESPONSE = 9,
  UNRECOGNIZED = -1,
}