// src/token-manager/HTTPClient/vehicle/protobuf/outputs/vcsec.ts
export enum OperationStatusE {
  OPERATIONSTATUS_OK = 0,
  OPERATIONSTATUS_WAIT = 1,
  OPERATIONSTATUS_ERROR = 2,
  UNRECOGNIZED = -1,
}
// ...
export enum VehicleLockStateE {
  VEHICLELOCKSTATE_UNLOCKED = 0,
  VEHICLELOCKSTATE_LOCKED = 1,
  VEHICLELOCKSTATE_INTERNAL_LOCKED = 2,
  VEHICLELOCKSTATE_SELECTIVE_UNLOCKED = 3,
  UNRECOGNIZED = -1,
}
// ...
export interface WhitelistInfo {
  numberOfEntries: number;
  whitelistEntries: KeyIdentifier[];
  slotMask: number;
}

export interface WhitelistEntryInfo {
  keyId: KeyIdentifier | undefined;
  publicKey: PublicKey | undefined;
  metadataForKey: KeyMetadata | undefined;
  slot: number;
  keyRole: Role;
}
// ...
export interface CommandStatus {
  operationStatus: OperationStatusE;
  signedMessageStatus?: SignedMessageStatus | undefined;
  whitelistOperationStatus?: WhitelistOperationStatus | undefined;
}

export interface UnsignedMessage {
  InformationRequest?: InformationRequest | undefined;
  RKEAction?: RKEActionE | undefined;
  closureMoveRequest?: ClosureMoveRequest | undefined;
  WhitelistOperation?: WhitelistOperation | undefined;
}

export interface ClosureStatuses {
  frontDriverDoor: ClosureStateE;
  frontPassengerDoor: ClosureStateE;
  rearDriverDoor: ClosureStateE;
  rearPassengerDoor: ClosureStateE;
  rearTrunk: ClosureStateE;
  frontTrunk: ClosureStateE;
  chargePort: ClosureStateE;
  tonneau: ClosureStateE;
}

export interface DetailedClosureStatus {
  tonneauPercentOpen: number;
}

export interface VehicleStatus {
  closureStatuses: ClosureStatuses | undefined;
  vehicleLockState: VehicleLockStateE;
  vehicleSleepStatus: VehicleSleepStatusE;
  userPresence: UserPresenceE;
  detailedClosureStatus: DetailedClosureStatus | undefined;
}

export interface FromVCSECMessage {
  vehicleStatus?: VehicleStatus | undefined;
  commandStatus?: CommandStatus | undefined;
  whitelistInfo?: WhitelistInfo | undefined;
  whitelistEntryInfo?: WhitelistEntryInfo | undefined;
  nominalError?: NominalError | undefined;
}
// ...
export const VehicleStatus: MessageFns<VehicleStatus> = {
  encode(message: VehicleStatus, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.closureStatuses !== undefined) {
      ClosureStatuses.encode(message.closureStatuses, writer.uint32(10).fork()).join();
    }
    if (message.vehicleLockState !== 0) {
      writer.uint32(16).int32(message.vehicleLockState);
    }
    if (message.vehicleSleepStatus !== 0) {
      writer.uint32(24).int32(message.vehicleSleepStatus);
    }
    if (message.userPresence !== 0) {
      writer.uint32(32).int32(message.userPresence);
    }
    if (message.detailedClosureStatus !== undefined) {
      DetailedClosureStatus.encode(message.detailedClosureStatus, writer.uint32(42).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): VehicleStatus {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseVehicleStatus();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.closureStatuses = ClosureStatuses.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.vehicleLockState = reader.int32() as any;
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.vehicleSleepStatus = reader.int32() as any;
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.userPresence = reader.int32() as any;
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.detailedClosureStatus = DetailedClosureStatus.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): VehicleStatus {
    return {
      closureStatuses: isSet(object.closureStatuses) ? ClosureStatuses.fromJSON(object.closureStatuses) : undefined,
      vehicleLockState: isSet(object.vehicleLockState) ? vehicleLockStateEFromJSON(object.vehicleLockState) : 0,
      vehicleSleepStatus: isSet(object.vehicleSleepStatus) ? vehicleSleepStatusEFromJSON(object.vehicleSleepStatus) : 0,
      userPresence: isSet(object.userPresence) ? userPresenceEFromJSON(object.userPresence) : 0,
      detailedClosureStatus: isSet(object.detailedClosureStatus)
        ? DetailedClosureStatus.fromJSON(object.detailedClosureStatus)
        : undefined,
    };
  },

  toJSON(message: VehicleStatus): unknown {
    const obj: any = {};
    if (message.closureStatuses !== undefined) {
      obj.closureStatuses = ClosureStatuses.toJSON(message.closureStatuses);
    }
    if (message.vehicleLockState !== 0) {
      obj.vehicleLockState = vehicleLockStateEToJSON(message.vehicleLockState);
    }
    if (message.vehicleSleepStatus !== 0) {
      obj.vehicleSleepStatus = vehicleSleepStatusEToJSON(message.vehicleSleepStatus);
    }
    if (message.userPresence !== 0) {
      obj.userPresence = userPresenceEToJSON(message.userPresence);
    }
    if (message.detailedClosureStatus !== undefined) {
      obj.detailedClosureStatus = DetailedClosureStatus.toJSON(message.detailedClosureStatus);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<VehicleStatus>, I>>(base?: I): VehicleStatus {
    return VehicleStatus.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<VehicleStatus>, I>>(object: I): VehicleStatus {
    const message = createBaseVehicleStatus();
    message.closureStatuses = (object.closureStatuses !== undefined && object.closureStatuses !== null)
      ? ClosureStatuses.fromPartial(object.closureStatuses)
      : undefined;
    message.vehicleLockState = object.vehicleLockState ?? 0;
    message.vehicleSleepStatus = object.vehicleSleepStatus ?? 0;
    message.userPresence = object.userPresence ?? 0;
    message.detailedClosureStatus =
      (object.detailedClosureStatus !== undefined && object.detailedClosureStatus !== null)
        ? DetailedClosureStatus.fromPartial(object.detailedClosureStatus)
        : undefined;
    return message;
  },
};

function createBaseFromVCSECMessage(): FromVCSECMessage {
  return {
    vehicleStatus: undefined,
    commandStatus: undefined,
    whitelistInfo: undefined,
    whitelistEntryInfo: undefined,
    nominalError: undefined,
  };
}

export const FromVCSECMessage: MessageFns<FromVCSECMessage> = {
  encode(message: FromVCSECMessage, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.vehicleStatus !== undefined) {
      VehicleStatus.encode(message.vehicleStatus, writer.uint32(10).fork()).join();
    }
    if (message.commandStatus !== undefined) {
      CommandStatus.encode(message.commandStatus, writer.uint32(34).fork()).join();
    }
    if (message.whitelistInfo !== undefined) {
      WhitelistInfo.encode(message.whitelistInfo, writer.uint32(130).fork()).join();
    }
    if (message.whitelistEntryInfo !== undefined) {
      WhitelistEntryInfo.encode(message.whitelistEntryInfo, writer.uint32(138).fork()).join();
    }
    if (message.nominalError !== undefined) {
      NominalError.encode(message.nominalError, writer.uint32(370).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FromVCSECMessage {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFromVCSECMessage();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.vehicleStatus = VehicleStatus.decode(reader, reader.uint32());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.commandStatus = CommandStatus.decode(reader, reader.uint32());
          continue;
        }
        case 16: {
          if (tag !== 130) {
            break;
          }

          message.whitelistInfo = WhitelistInfo.decode(reader, reader.uint32());
          continue;
        }
        case 17: {
          if (tag !== 138) {
            break;
          }

          message.whitelistEntryInfo = WhitelistEntryInfo.decode(reader, reader.uint32());
          continue;
        }
        case 46: {
          if (tag !== 370) {
            break;
          }

          message.nominalError = NominalError.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FromVCSECMessage {
    return {
      vehicleStatus: isSet(object.vehicleStatus) ? VehicleStatus.fromJSON(object.vehicleStatus) : undefined,
      commandStatus: isSet(object.commandStatus) ? CommandStatus.fromJSON(object.commandStatus) : undefined,
      whitelistInfo: isSet(object.whitelistInfo) ? WhitelistInfo.fromJSON(object.whitelistInfo) : undefined,
      whitelistEntryInfo: isSet(object.whitelistEntryInfo)
        ? WhitelistEntryInfo.fromJSON(object.whitelistEntryInfo)
        : undefined,
      nominalError: isSet(object.nominalError) ? NominalError.fromJSON(object.nominalError) : undefined,
    };
  },

  toJSON(message: FromVCSECMessage): unknown {
    const obj: any = {};
    if (message.vehicleStatus !== undefined) {
      obj.vehicleStatus = VehicleStatus.toJSON(message.vehicleStatus);
    }
    if (message.commandStatus !== undefined) {
      obj.commandStatus = CommandStatus.toJSON(message.commandStatus);
    }
    if (message.whitelistInfo !== undefined) {
      obj.whitelistInfo = WhitelistInfo.toJSON(message.whitelistInfo);
    }
    if (message.whitelistEntryInfo !== undefined) {
      obj.whitelistEntryInfo = WhitelistEntryInfo.toJSON(message.whitelistEntryInfo);
    }
    if (message.nominalError !== undefined) {
      obj.nominalError = NominalError.toJSON(message.nominalError);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<FromVCSECMessage>, I>>(base?: I): FromVCSECMessage {
    return FromVCSECMessage.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<FromVCSECMessage>, I>>(object: I): FromVCSECMessage {
    const message = createBaseFromVCSECMessage();
    message.vehicleStatus = (object.vehicleStatus !== undefined && object.vehicleStatus !== null)
      ? VehicleStatus.fromPartial(object.vehicleStatus)
      : undefined;
    message.commandStatus = (object.commandStatus !== undefined && object.commandStatus !== null)
      ? CommandStatus.fromPartial(object.commandStatus)
      : undefined;
    message.whitelistInfo = (object.whitelistInfo !== undefined && object.whitelistInfo !== null)
      ? WhitelistInfo.fromPartial(object.whitelistInfo)
      : undefined;
    message.whitelistEntryInfo = (object.whitelistEntryInfo !== undefined && object.whitelistEntryInfo !== null)
      ? WhitelistEntryInfo.fromPartial(object.whitelistEntryInfo)
      : undefined;
    message.nominalError = (object.nominalError !== undefined && object.nominalError !== null)
      ? NominalError.fromPartial(object.nominalError)
      : undefined;
    return message;
  },
};
// ...
export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}